BaseCollectionView.setFilter() would not call the callback if the filter hadn't changed.
Fixes a bug where a TableView or ListView might never seem to finish loading if you were setting a filter on it.

Also cleaned up a bunch of Generics stuff to remove warnings.

Also changed a bunch of methods to take an AsyncCallbackGroup so that they don't have to create their own group all the time.